rm the old query editor and datasource
