--- src/datasource/query_editor.html
@@ -1,113 +1,0 @@
-<query-editor-row query-ctrl="ctrl" can-collapse="false">
-  <div class="gf-form-inline">
-    <div class="gf-form">
-      <label class="gf-form-label width-7">Device</label>
-      <metric-segment
-        class="width-11"
-        segment="ctrl.deviceSegment"
-        get-options="ctrl.getDevices($query)"
-        on-change="ctrl.onDeviceChange()"
-      ></metric-segment>
-    </div>
-    <div class="gf-form">
-      <label class="gf-form-label width-7">Site</label>
-      <metric-segment
-        class="width-11"
-        segment="ctrl.siteSegment"
-        get-options="ctrl.getSites($query)"
-        on-change="ctrl.onSiteChange()"
-      ></metric-segment>
-    </div>
-    <div class="gf-form gf-form--grow">
-      <div class="gf-form-label gf-form-label--grow"></div>
-    </div>
-  </div>
-  <div class="gf-form-inline">
-    <div class="gf-form">
-      <label class="gf-form-label width-7">Metric</label>
-      <metric-segment
-        class="width-11"
-        segment="ctrl.metricSegment"
-        get-options="ctrl.getMetrics($query)"
-        on-change="ctrl.onMetricChange()"
-      ></metric-segment>
-    </div>
-    <div class="gf-form">
-      <label class="gf-form-label width-7">Unit</label>
-      <metric-segment
-        class="width-11"
-        segment="ctrl.unitSegment"
-        get-options="ctrl.getUnits($query)"
-        on-change="ctrl.onUnitChange()"
-      ></metric-segment>
-    </div>
-    <div class="gf-form">
-      <label class="gf-form-label width-7">Data mode</label>
-      <div class="gf-form-select-wrapper width-11">
-        <select class="gf-form-input" ng-model="ctrl.target.mode" ng-options="f.value as f.text for f in ctrl.queryModes" ng-change="ctrl.refresh()"></select>
-      </div>
-    </div>
-    <div class="gf-form gf-form--grow">
-      <div class="gf-form-label gf-form-label--grow"></div>
-    </div>
-  </div>
-  <div class="gf-form-inline">
-    <div class="gf-form">
-        <label class="gf-form-label width-7">DNS Lookup</label>
-        <metric-segment
-          class="width-11"
-          segment="ctrl.hostnameLookup"
-          get-options="ctrl.getHostnameLookupOptionValues()"
-          on-change="ctrl.onHostnameLookupChange()"
-        ></metric-segment>
-    </div>
-
-    <div class="gf-form">
-      <label class="gf-form-label width-7">Prefix</label>
-      <input type="text" class="gf-form-input width-11" ng-model='ctrl.target.prefix' ng-blur="ctrl.onPrefixChange()" />
-    </div>
-    <div class="gf-form gf-form--grow">
-        <div class="gf-form-label gf-form-label--grow"></div>
-      </div>
-  </div>
-  <div class="gf-form-inline">
-    <div class="gf-form">
-      <label class="gf-form-label width-7">Filters</label>
-    </div>
-    <span class="pointer gf-form-label width-2" ng-click="ctrl.addFilter()">
-      <i class="fa fa-plus"></i>
-    </span>
-    <div class="gf-form">
-      <metric-segment
-        ng-if="ctrl.conjunctionSegment !== undefined"
-        class="width-3"
-        segment="ctrl.conjunctionSegment"
-        get-options="ctrl.getConditionsOptionValues()"
-        on-change="ctrl.onConjunctionSegmentChange()"
-      ></metric-segment>
-    </div>
-    <div class="gf-form gf-form--grow">
-      <div class="gf-form-label gf-form-label--grow"></div>
-    </div>
-  </div>
-  <div ng-repeat="filter in ctrl.filterList" class="gf-form-inline">
-    <div class="gf-form">
-      <label class="gf-form-label width-7"></label>
-    </div>
-    <div class="gf-form">
-      <metric-segment class="width-8" segment="filter.keySegment" get-options="ctrl.getFilterOptionKeys()"
-        on-change="ctrl.onFilterInputChange($index, 'keySegment')"></metric-segment>
-      <metric-segment ng-if="filter.operatorSegment !== undefined" class="width-2" segment="filter.operatorSegment" get-options="ctrl.getOperatorOptionValues()"
-        on-change="ctrl.onFilterInputChange($index, 'operatorSegment')"></metric-segment>
-      <metric-segment ng-if="filter.valueSegment !== undefined" class="width-8" segment="filter.valueSegment" get-options="ctrl.getFilterOptionValues($index)"
-        on-change="ctrl.onFilterInputChange($index, 'valueSegment')"></metric-segment>
-      <label ng-if="filter.conjunctionOperator !== undefined" class="gf-form-label width-3">{{ filter.conjunctionOperator }}</label>
-      <span class="pointer gf-form-label width-2" ng-click="ctrl.deleteFilter($index)">
-        <i class="fa fa-times"></i>
-      </span>
-    </div>
-    <div class="gf-form gf-form--grow">
-      <div class="gf-form-label gf-form-label--grow"></div>
-    </div>
-  </div>
-</query-editor-row>
